Enclose two systables in the main system table object
st->current is the system table which gets loaded from a file on startup
and is then put into use.

st->new is the new system table which is built from decoded System
Table LPDUs. When the new version is decoded successfully and is
newer than the current one, it will replace the current one and will
(optionally) get stored in a file for future use.