Revert AC cache TTL to 3600 seconds
Shortened TTL purges many mappings which are still valid and it
reduces the amount of mismatches by a negligible amount. This
problem seems to be mainly caused by misconfigured HFDL radios
on aircraft, where the hex configured on the radio does not match
the hex set on the ADS-B transponder.